create new file AssetsItem
